Pull out logfile locating logic.
This moves a few hundred lines of 'where-to-put-the-log-file' logic
from the LoggingManager into a new class.

Everything is functionally identical for now, except the Locator
uses a ConfigBinder for its settings.
